HttpUtils is no longer a local static class but a global utility class
