+++ app/src/main/java/com/twofauth/android/main_activity/CheckForAppUpdates.java
@@ -1,0 +1,4 @@
+package com.twofauth.android.main_activity;
+
+public class CheckForAppUpdates {
+}
